Refactor bindFragmentChildren to align with other binding
